// src/app/models/destino.ts
export interface Destino{
    _id?: string
    nombre: string
    telefono: string
    email: string
    usuarioRed: string
    filial: string
    dominio: string
    subdominio: string
    proceso: string
    codigoMac: string
    ambiente: string
    usuarioCnx: string
    protocoloTransmision:string
    puerto: string
    tipoSvr: string
    Servidor: string
    usuarioConexionDestino: string
    metodoAutenticacion: string
    UbicacionServidor: string
    rutaEntrega: string
    createdAt?: string
    updatedAt?: string
}